requirements.txt
alembic==1.6.5
appdirs==1.4.4
asgiref==3.3.1
astroid==2.6.2
bcrypt==3.2.0
beautifulsoup4==4.9.3
black==21.5b1
certifi==2020.11.8
cffi==1.14.5
chardet==3.0.4
click==7.1.2
colorama==0.4.4
distlib==0.3.1
Django==3.1.3
djangorestframework==3.12.4
filelock==3.0.12
flake8==3.9.2
Flask==1.1.4
flask-marshmallow==0.14.0
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
html-parser==0.2
idna==2.10
inflect==5.3.0
isort==5.9.1
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
luhn==0.2.0
lxml==4.6.1
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.11.1
marshmallow-sqlalchemy==0.25.0
mccabe==0.6.1
mergedeep==1.3.4
mutagen==1.45.1
mypy-extensions==0.4.3
numpy==1.19.3
pandas==1.1.4
pathspec==0.8.1
peewee==3.13.3
ply==3.11
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pycparser==2.20
pyenv-win==2.64.6.1
pyflakes==2.3.1
PyJWT==2.1.0
pylint==2.9.3
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
python-json-logger==2.0.1
pytz==2020.1
regex==2021.4.4
requests==2.24.0
sentry-sdk==0.19.4
six==1.15.0
soupsieve==2.0.1
spotify-dl==7.6.0
spotipy==2.16.1
sqlacodegen==2.3.0
SQLAlchemy==1.3.12
sqlparse==0.4.1
toml==0.10.2
urllib3==1.25.11
virtualenv==20.4.4
Werkzeug==1.0.1
wrapt==1.12.1
youtube-dl==2021.5.16
